pyproject.toml
[project]
name = "autofilter"
version = "0.1.0"
description = "Fine-tune LLMs to convert natural language queries into structured filter expressions."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "trl==0.21.0",
    "transformers==4.55.0",
    "peft==0.17.1",
    "datasets==4.0.0",
    "accelerate>=1.12.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "typer>=0.15.0",
    "wandb>=0.19.0",
    "bitsandbytes>=0.43.0",
    "tqdm>=4.60.0",
]